Move out interesting logic from Top-level CLI We move it into a separate module for testing.
